PEP8 Rename mixin modules
